Separate SerializedBlendShapeClipEditor
