+++ Scripts/BlendShape/Editor/SerializedBlendShapeClipEditor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: eaafa3ff7bf991642b922e6af7ecbbc0
+timeCreated: 1541081003
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
